com- debug crud for product

+++ src/main/java/com/digipay/inventory/service/productService/ProductServiceImpl.java
@@ -9,6 +9,8 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
 
 @Service
@@ -19,7 +21,7 @@ public class ProductServiceImpl extends MajorService implements ProductService {
     private CheckingImpl checking;
 
     @Autowired
-    public ProductServiceImpl(ProductRepository productRepository,CheckingImpl checking) {
+    public ProductServiceImpl(ProductRepository productRepository, CheckingImpl checking) {
         this.productRepository = productRepository;
         this.checking = checking;
     }
@@ -44,8 +46,18 @@ public class ProductServiceImpl extends MajorService implements ProductService {
     }
 
     @Override
-    public void deleteById(Long id) {
-        productRepository.deleteById(id);
+    public Boolean deleteById(Long id) throws Exception {
+        Product product = new Product();
+        try {
+            product = productRepository.getProductById(id);
+            if (Objects.nonNull(product)) {
+                productRepository.deleteById(id);
+                return true;
+            } else
+                return false;
+        } catch (Exception e) {
+            throw new Exception(e.getMessage());
+        }
     }
